Fix incorrect second reference to COMMAND for bun-node-headers
Run `ninja -C build/debug -t commands bun-node-headers` before and after
this commit to see the difference. Before this commit, the two commands
were mashed together with a stray "cmake" argument in between. That
fails an assert on my CMake from Gentoo, for some reason. On the Debian
and official build of CMake, it works fine. Still, this doesn't look
intended, considering `&&` is used in numerous other invocations of
register_command, so here's a fix anyway.