Add a minimum viable implementation for CompressionStream for zlib algos
The code for CompressionStreamEncoder and DecompressionStreamDecoder are
present in the same compression.zig file, and they are both instances of
a generic type since they both share very similar methods, just with
different contexts for compression/decompression. While the code for
zlib is similar regardless of whether compression/decompression is done,
I'm anticipating it is not the same for Brotli or Zstd.

Is this the most efficient implementation? Definitely not.
Is it fully compatible with Node.js? Not quite...errors are different.